Fixed displaying appointments on the main page
In case an appointment is made and a patient or a doctor linked
to that appointment is later deleted, the appointment remains in the
database (with the ID of the patient or the doctor). This is the same
behaviour as in the Clinic application that I previously written in
Java and I wanted to keep that behaviour in Clinic_Kotlin_Gradle as
well. However, the solution used in the Java application did not work
this time - I was getting an EntityNotFoundException. The proposed
solution involves adding @NotFound(action = NotFoundAction.IGNORE)
annotations in the Appointment class and changing the filtering method
getAllForExistingPatientsAndDoctors() in the AppointmentService class.
This way, the application at this stage works and looks exactly the
same and offers the same functionality as the previous Java application.